Remove config file
I originally experimented with this file because during the workshop
students freaked out when Streamlit asked for their email address when first running the app.

I thought that setting "prompt = false" would obviate that. But it
turns out it just suppresses the prompt, but then gives an error because
it still needs an email.

I then tried to set a placeholder email. While that might work, research shows that Streamlit might use this email for deployment to Streamlit Cloud. So it might cause problems in later lessons.

While not prompting for email is certainly smoother for the students, the fact is that I'm about to launch the course, and this isn't worth causing more errors for the students. I'd rather update the instructions to warn students about the email request and explain that it's tied to deployment, which they'll do later.