Better fix for suppressing Streamlit email prompt
Based on https://discuss.streamlit.io/t/disabling-email-prompt-when-students-first-run-streamlit/119247/2

It seems like my original fix of using .streamlit/config.toml to set
showEmailPrompt=false was correct. The reason why it didn't work seems to have been a bug in Streamlit. On the forum someone recommended I update the version of Streamlit I'm using.

So this updates Streamlit from 1.48.0 to 1.49.1 and adds back in the config.toml file.